Explanatory VueTS code behind file
